Fix for when CLASSPATH environment path is set
Say, I had a CLASSPATH of:
export CLASSPATH=opsin-big-0.1.0.jar
rcdk would concatenate that directly to form a classpath of
opsin-big-0.1.0.jarfilelocation/../java/lib/cdk-1.2 ...
giving no colon separation in the path.

This just puts that colon there.
In the case where the classpath is empty, there's an extra colon, but I
don't think that matters.